Increase repository priority for ceph-hammer
With CentOS 7.3, we need to increase priority of ceph
Storage SIG priority as some packages conflicts with
same NVR in base repo.

https://review.openstack.org/#/c/410823/ implements
the same change in puppet-ceph, but in weirdo we
dont't use ceph::repo to configure the repo, so it's
not being applied.

Base images in review.r.o include librbd1 from base
repo. This patch reinstalls it after enabling ceph
repo if it's installed.

Change-Id: Iffb7feed14e383fcfafe7ae36b6d081bf2c5b28d