Set ZUUL_BRANCH variable to version
In weirdo we use "version" variable for the branch of p-o-i
to use for the jobs. This is working fine unless there is zuul
installed in the instances where jobs are running. In this case
zuul-clone is used to download p-o-i and it's always using master
instead of the required branch.

This patch is setting ZUUL_BRANCH variable to version, so that
zuul clones the right branch.

Change-Id: Ia6a99bafebdd0772d78bd0dd8fcd437e9b2cbcd8