Remove zuul-cloner executable before running p-o-i run_test.sh
In [1] p-o-i checks existence of /usr/zuul-env/bin/zuul-cloner to decide
how to install puppet modules. When running weirdo in jenkins jobs or in
other's gates, as for example in rdoinfo, this is breaking as it allways
clones p-o-i from master, even when version variable points to
stable/<release>. As weirdo already has the logic to clone p-o-i from
the right branch, let's let it do the clone and run_test.sh simply copy
it to the right location (what does in install_all).

[1] https://github.com/openstack/puppet-openstack-integration/blob/4bbea36c6f7008a194245400bb29e957f10f17c1/functions#L65

Change-Id: Ia6a99bafebdd0772d78bd0dd8fcd437e9b2cbcd8