Remove "static" field while including tasks
Instead, we use "include_tasks" to tell tasks must be
included dynamically. The use of 'static' has been
deprecated and will be removed in version 2.12.

Plus, we force evaluate 'enable_puppet_modules_rpm' as
a boolean and not a bare variable, as this behaviour
will be removed in version 2.12.

Finally, we remove Jinja2 templating in some conditional
statements.

Change-Id: If01c97913557e0dbee927e9f0167e1515217359c