Symlink p-o-i repo to the path where it expects
p-o-i has a hardcoded path where it expects the upstream repo. As we are
cloning into a different one, it's cloning it from master (as
ZUUL_BRANCH is undefined)[1].

We may define ZUUL_BRANCH to version variable but then we would break
the feature that weirdo has of testing an ongoing patch of p-o-i using
variable `refspec`.

[1] https://github.com/openstack/puppet-openstack-integration/blob/f384effdb42e0cbe19e137f0411b09ac8fef5b48/functions#L68-L72

Change-Id: Ice6d204cc5d32db85b5d2522238e777361ac3040